pyproject.toml
[project]
name = "realtimemap"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "fastapi-babel>=1.0.0",
    "fastapi-pagination>=0.13.1",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "geoalchemy2>=0.17.1",
    "geojson-pydantic>=2.0.0",
    "orjson>=3.10.18",
    "pydantic-settings>=2.9.1",
    "shapely>=2.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.10",
]
